+++ src/main/kotlin/lotto/service/issue/LottoIssueService.kt
@@ -1,4 +1,4 @@
-package lotto.service
+package lotto.service.issue
 
 import lotto.domain.Lotto
